CSP

Description

Documentation

class Assignment a => CSP a v d | a -> v, a -> d where

Represents the interfaces needed by the current Solver so that the recursive backtracking implementation can work using MRV and LCV

Minimal complete definition

legalValues, relatedVariables, addAssignment, minValueCount

Methods

legalValues :: v -> a -> Set d

relatedVariables :: v -> a -> [v]

addAssignment :: v -> d -> a -> a

minValueCount :: a -> v

class Assignment a where

Represents the interface required by CSP assignments so that it can be checked that it is valid but unsolved and valid and solved, respectively

Minimal complete definition

consistent, complete

Methods

consistent :: a -> Bool

complete :: a -> Bool