SudokuBoard

Description

Documentation

data SudokuBoard

Contains the state of the board.

Constructors

Board (Seq (Seq SudokuDigit))

type BoardPosition = (Int, Int)

Represents the row, column pairs to index positions in the board.

getAllRelatedDigits :: BoardPosition -> SudokuBoard -> Seq SudokuDigit

Returns a sequence of all the digits in the same rowcolcage as the given position, including the given position.

getAllRelatedPositions :: BoardPosition -> [BoardPosition]

Returns a list of all the board positions in the same rowcolcage as the given position, including the given position.

getDigit :: BoardPosition -> SudokuBoard -> SudokuDigit

Returns the digit in the given board at the given position.

strToBoard :: String -> SudokuBoard

Converts a string into a sudoku board.

Expects a string of 81 characters of digits 1-9. Any other char is interpreted as a Blank.

updateBoard :: BoardPosition -> SudokuDigit -> SudokuBoard -> SudokuBoard

Returns a new board with the digit in the given position.

validBoard :: SudokuBoard -> Bool

Returns a boolean indicating if the board is valid(no duplicates in any rowcolcage, blanks allowed).

solvedBoard :: SudokuBoard -> Bool

Returns a boolean indicating if the board is solved(no duplicates in any rowcolcage, blanks not allowed).