SudokuDigit

Description

Documentation

data SudokuDigit

Covers the available assignments for a SudokuBoard.

Used instead of Ints to limit errors when creating boards.

Constructors

Blank

One

Two

Three

Four

Five

Six

Seven

Eight

Nine

charToDigit :: Char -> SudokuDigit

Converts Char to SudokuDigit for encoding Strings to SudokuBoards.

sudokuDomain :: Set SudokuDigit

The domain of values that are legal to place in a SudokuBoard. Excludes Blank, as this is a placeholder for unsolved positions.